Move crud.Store out of the internal package
Now that the wrappers are gone, this is the entry point of the package, so it has to be public
